+++ src/store/src/Bridge/AzureSearch/CHANGELOG.md
@@ -8,6 +8,7 @@ CHANGELOG
  * [BC BREAK] The `endpointUrl` parameter for `SearchStore` has been removed
  * [BC BREAK] The `apiKey` parameter for `SearchStore` has been removed
  * [BC BREAK] The `apiVersion` parameter for `SearchStore` has been removed
+ * Add `count` method
 
 0.4
 ---