+++ src/store/src/Bridge/Pinecone/CHANGELOG.md
@@ -4,7 +4,9 @@ CHANGELOG
 0.7
 ---
 
+ * Add `count` method
  * Add support for `includeMetadata` and `includeValues` options
+ * Add `count` method
 
 0.1
 ---